+++ rollup.config.mjs
@@ -1,4 +1,5 @@
 import commonjs from "@rollup/plugin-commonjs";
+// eslint-disable-next-line import/no-named-as-default
 import nodeResolve from "@rollup/plugin-node-resolve";
 import external from "rollup-plugin-peer-deps-external";
 import terser from "@rollup/plugin-terser";
@@ -41,7 +42,6 @@ export default [
       sourcemap: true,
       globals: { react: "React" },
       exports: "named",
-      sourcemap: true,
       interop: "auto",
     },
     plugins,
@@ -55,7 +55,6 @@ export default [
       sourcemap: true,
       globals: { react: "React" },
       exports: "named",
-      sourcemap: true,
     },
     plugins,
   },
@@ -68,7 +67,6 @@ export default [
       sourcemap: true,
       globals: { react: "React" },
       exports: "named",
-      sourcemap: true,
     },
     plugins,
   },