+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- migrated to eslint 9 flat config
+- removed obsolete storybook
+
 ## [5.13.0] - 2025-10-08
 
 ### Changed